+++ packages/patrol/android/src/main/kotlin/pl/leancode/patrol/AutomatorServer.kt
@@ -168,16 +168,24 @@ class AutomatorServer(private val automation: Automator) : NativeAutomatorServer
 
     override fun tap(request: TapRequest) {
         if (request.selector != null) {
+            // Remove instance before creating bySelector, as it's not supported
+            var selector2 = request.selector.copy(instance = null)
+            val bySelector = selector2.toBySelector()
+
             automation.tap(
                 uiSelector = request.selector.toUiSelector(),
-                bySelector = request.selector.toBySelector(),
+                bySelector = bySelector,
                 index = request.selector.instance?.toInt() ?: 0,
                 timeout = request.timeoutMillis
             )
         } else if (request.androidSelector != null) {
+            // Remove instance before creating bySelector, as it's not supported
+            var androidSelector2 = request.androidSelector.copy(instance = null)
+            val bySelector = androidSelector2.toBySelector()
+
             automation.tap(
                 uiSelector = request.androidSelector.toUiSelector(),
-                bySelector = request.androidSelector.toBySelector(),
+                bySelector = bySelector,
                 index = request.androidSelector.instance?.toInt() ?: 0,
                 timeout = request.timeoutMillis
             )